# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NutriPlan API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile and goals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── FOODS ─────────────────────────────────────────────────────────────────
...
  /meals:
    get:
      operationId: listMeals
      tags: [meals]
      summary: List user's saved meals
      responses:
        "200":
          description: Meals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
    post:
      operationId: createMeal
      tags: [meals]
      summary: Create a saved meal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /meals/{id}:
    get:
      operationId: getMeal
      tags: [meals]
      summary: Get a meal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMeal
      tags: [meals]
      summary: Update a saved meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMeal
      tags: [meals]
      summary: Delete a saved meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /meals/{id}/log:
    post:
      operationId: logMeal
      tags: [meals]
      summary: Log an entire saved meal to a date/mealType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealLogInput"
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodLog"

  # ── MEAL PLANS ────────────────────────────────────────────────────────────
  /meal-plans:
    get:
      operationId: listMealPlans
      tags: [mealPlans]
      summary: Get meal plan entries for a week
      parameters:
        - name: weekStart
          in: query
          required: true
          schema:
            type: string
            description: "Week start date in YYYY-MM-DD format (Monday)"
      responses:
        "200":
          description: Meal plan entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealPlanEntry"
    post:
      operationId: createMealPlanEntry
      tags: [mealPlans]
      summary: Add a meal to the weekly plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealPlanEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealPlanEntry"
  /meal-plans/{id}:
    delete:
      operationId: deleteMealPlanEntry
      tags: [mealPlans]
      summary: Remove a meal from the weekly plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      required: [id, email, name, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        activityLevel:
          type: ["string", "null"]
        dailyCalorieGoal:
          type: ["integer", "null"]
        dailyProteinGoal:
          type: ["number", "null"]
        dailyCarbsGoal:
          type: ["number", "null"]
        dailyFatGoal:
          type: ["number", "null"]
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        age:
          type: ["integer", "null"]
        gender:
          type: ["string", "null"]
        weight:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        activityLevel:
          type: ["string", "null"]
        dailyCalorieGoal:
          type: ["integer", "null"]
        dailyProteinGoal:
          type: ["number", "null"]
        dailyCarbsGoal:
          type: ["number", "null"]
        dailyFatGoal:
          type: ["number", "null"]

    # ── FOOD SCHEMAS ─────────────────────────────────────────────────────────
    Food:
      type: object
      required: [id, name, calories, protein, carbs, fat, servingSize, isCustom]
      properties:
        id:
          type: integer
        name:
          type: string
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        fiber:
          type: ["number", "null"]
        sugar:
          type: ["number", "null"]
        sodium:
          type: ["number", "null"]
        servingSize:
          type: string
        servingUnit:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        isCustom:
          type: boolean
        userId:
          type: ["integer", "null"]
...
    FoodLog:
      type: object
      required: [id, userId, date, mealType, quantity, calories, protein, carbs, fat, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        foodId:
          type: ["integer", "null"]
        food:
          $ref: "#/components/schemas/Food"
        customFoodName:
          type: ["string", "null"]
        quantity:
          type: number
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    MealItem:
      type: object
      required: [id, foodId, quantity, calories, protein, carbs, fat]
      properties:
        id:
          type: integer
        foodId:
          type: integer
        food:
          $ref: "#/components/schemas/Food"
        quantity:
          type: number
        calories:
          type: integer
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
    MealItemInput:
      type: object
      required: [foodId, quantity, calories, protein, carbs, fat]
      properties:
        foodId:
          type: integer
        quantity:
          type: number
          minimum: 0
        calories:
          type: integer
          minimum: 0
        protein:
          type: number
          minimum: 0
        carbs:
          type: number
          minimum: 0
        fat:
          type: number
          minimum: 0
    Meal:
      type: object
      required: [id, userId, name, totalCalories, totalProtein, totalCarbs, totalFat, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        totalCalories:
          type: integer
        totalProtein:
          type: number
        totalCarbs:
          type: number
        totalFat:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/MealItem"
        createdAt:
          type: string
    MealInput:
      type: object
      required: [name, items]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/MealItemInput"
    MealUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/MealItemInput"
    MealLogInput:
      type: object
      required: [date, mealType]
      properties:
        date:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]

    # ── MEAL PLAN SCHEMAS ─────────────────────────────────────────────────────
    MealPlanEntry:
      type: object
      required: [id, userId, date, mealType, mealId]
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        mealId:
          type: integer
        meal:
          $ref: "#/components/schemas/Meal"
    MealPlanEntryInput:
      type: object
      required: [date, mealType, mealId]
      properties:
        date:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        mealId:
          type: integer

    # ── DASHBOARD SCHEMAS ─────────────────────────────────────────────────────
...