# lib/api-zod/src/generated/types/mealLogInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriPlan API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MealLogInputMealType } from './mealLogInputMealType';

export interface MealLogInput {
  date: string;
  mealType: MealLogInputMealType;
}