# lib/api-zod/src/generated/types/mealPlanEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NutriPlan API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Meal } from './meal';
import type { MealPlanEntryMealType } from './mealPlanEntryMealType';

export interface MealPlanEntry {
  id: number;
  userId: number;
  date: string;
  mealType: MealPlanEntryMealType;
  mealId: number;
  meal?: Meal;
}